Various compilation issues resolved
Could not get this to compile, Im guessing the language update of
recent broke this project

+++ AudioKit/Platforms/OSX/CsoundObj.m
@@ -315,7 +315,7 @@ static void messageCallback(CSOUND *cs, int attr, const char *format, va_list va
         return nil;
     }
     CSOUND *csound = [self getCsound];
-    float *spout = csoundGetSpout(csound);
+    double *spout = csoundGetSpout(csound);
     int nchnls = csoundGetNchnls(csound);
     int ksmps = csoundGetKsmps(csound);
     return [NSMutableData dataWithBytes:spout length:(nchnls * ksmps * sizeof(MYFLT))];
@@ -339,7 +339,7 @@ static void messageCallback(CSOUND *cs, int attr, const char *format, va_list va
         return nil;
     }
     CSOUND *csound = [self getCsound];
-    float *spin = csoundGetSpin(csound);
+    double *spin = csoundGetSpin(csound);
     int nchnls = csoundGetNchnls(csound);
     int ksmps = csoundGetKsmps(csound);
     return [NSMutableData dataWithBytes:spin length:(nchnls * ksmps * sizeof(MYFLT))];
